requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
backoff==2.2.1
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
comm==0.2.2
cryptography==42.0.8
datasets==2.14.7
debugpy==1.8.1
decorator==5.1.1
dill==0.3.7
distro==1.9.0
dspy-ai==2.4.9
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2023.10.0
grpcio==1.64.1
grpcio-health-checking==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.2
jupyter_core==5.7.2
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.0.5
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.31.0
optuna==3.6.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
protobuf==5.27.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
ruff==0.4.8
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
structlog==24.2.0
tomli==2.0.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
validators==0.28.3
wcwidth==0.2.13
weaviate-client==4.6.4
xxhash==3.4.1
yarl==1.9.4
